Split ModSettingChangeTracker into separate file
